rename controller to be consistent
