+++ server/controllers/currency.controller.ts
@@ -1,5 +1,5 @@
 import Currency from '../models/currency.model';
-import BaseCtrl from './base';
+import BaseCtrl from './base.controller';
 
 export default class CurrencyCtrl extends BaseCtrl {
   model: any;
