+++ server/controllers/exchange.controller.ts
@@ -1,5 +1,5 @@
 import Exchange from '../models/exchange.model';
-import BaseCtrl from './base';
+import BaseCtrl from './base.controller';
 
 export default class ExchangeCtrl extends BaseCtrl {
   model: any;
