+++ server/controllers/role.controller.ts
@@ -1,9 +1,9 @@
 import Role from '../models/role.model';
-import BaseCtrl from './base';
+import BaseCtrl from './base.controller';
 
 export default class RoleCtrl extends BaseCtrl {
   model: any;
-  
+
   constructor() {
     super();
     this.model = Role;
